+++ include/gw2formats/pf/helpers/RefList.h
@@ -63,7 +63,7 @@ namespace gw2f {
 					return *this;
 				}
 
-				const T& operator[]( uint32 p_index ) const {
+				const T& operator[]( size_t p_index ) const {
 					if ( p_index >= m_data->size( ) ) { throw std::out_of_range( "The given index is too large." ); }
 					return ( *m_data )[p_index];
 				}
